First commit
As you can see I’ve been working on this for a little while before
placing it here. I wanted to get most of the initial functionality in.
It’s rather hackish at the moment, lot’s of room for improvement, but
it’s a good start. Read the readme for some further details.