Introduced dual column layout and template caching
I held off on the dual column layout as it meant I had to write a
template function to implement it and I wasn’t too keen on figuring
that out, but after looking into it online it’s actually not as hard as
I was thinking as you can see here in the code.

Likewise, I implemented template caching to further improve
performance.  Both of these features are possible thanks to the go
tutorials, documentation and of course, stackoverflow.